fix: address review — material leak ExpansionRealWorldBridge, null label, CompatAll, LayerMask cast
Agent-Logs-Url: https://github.com/DiosDeJuan/XINO/sessions/e6d50a25-6734-4f3a-951c-6d569c998eac

+++ Animo/Assets/StoreSimulator/Scripts/DeliverySystem.cs
@@ -157,7 +157,7 @@ namespace FLOBUK.StoreSimulator
 
             // Safe layer mask: use InteractionSystem if available, otherwise use default (all layers).
             int layerMask = InteractionSystem.Instance != null
-                ? (int)InteractionSystem.Instance.layerMask
+                ? InteractionSystem.Instance.layerMask
                 : Physics.DefaultRaycastLayers;
 
             //starting from the deliveryStart position, do a raycast until the end of deliveryDirection to find the lowest
